fix return type of get_seconds

+++ src/modules/data.h
@@ -2,7 +2,7 @@
 #define DATA_H
 
 struct tm *get_time();
-double get_seconds();
+int get_seconds();
 double get_cpu_utilization(char *cpuN);
 double get_memory_usage();
